requirements.txt
flask==2.0.1
numpy==1.21.2
scikit-learn==0.24.2
sentence-transformers==2.2.2
torch==1.9.0
python-dotenv==0.19.0
Werkzeug==2.0.1
requests==2.26.0
PyPDF2==3.0.1
python-docx==0.8.11
python-pptx>=0.6.21
openpyxl>=3.1.2
huggingface-hub==0.8.1
urllib3==1.26.15
chardet==4.0.0
charset-normalizer==2.0.12
pandas>=2.0.0
xlrd>=2.0.1
python-magic>=0.4.27
tqdm>=4.65.0
faiss-cpu>=1.7.4
jieba>=0.42.1
redis>=4.5.4
itchat-uos>=1.5.0.dev0
wxpy>=0.3.9
asyncio>=3.4.3
aiohttp>=3.8.4
loguru>=0.7.0 